azure kms integration test added
